requirements.txt
absl-py==0.12.0
aiofiles==0.6.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.7.4
boto3==1.17.45
botocore==1.20.45
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
datapackage==1.15.2
dill==0.3.3
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ijson==3.1.4
importlib-metadata==3.10.0
isodate==0.6.0
jieba3k==0.35.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonlines==2.0.0
jsonpointer==2.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
linear-tsv==1.1.0
loader==2017.9.11
Logbook==1.5.3
logcc==2018.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
multidict==5.1.0
multiprocess==0.70.11.1
murmurhash==1.0.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pathy==0.4.0
pendulum==2.1.2
Pillow==8.2.0
pony==0.7.14
preshed==3.0.5
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.3.6
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
sec-edgar-downloader==4.0.2
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.2
SQLAlchemy==1.4.5
srsly==2.4.0
tableschema==1.20.2
tabulator==1.53.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcc==2019.8.2
termcolor==1.1.0
terminaltables==3.1.0
thinc==8.0.2
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.4.2
typer==0.3.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
wasabi==0.8.2
websockets==8.1
Werkzeug==1.0.1
wrap==2018.3.1
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
zipp==3.4.1
